+++ example/lib/research_package_objects/infomed_consent_objects.dart
@@ -8,14 +8,12 @@ RPConsentSection overviewSection = RPConsentSection(
 
 RPConsentSection dataGatheringSection = RPConsentSection(
     RPConsentSectionType.DataGathering)
-  ..summary =
-      "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Aliquam ut aliquet augue, sed maximus sapien. Cras non sem vulputate, tincidunt est in, euismod orci. Aenean elit nisl, convallis vel lacus eget, pellentesque laoreet nunc. Proin sed turpis odio. Aliquam quis metus eget tortor tristique posuere sed blandit velit. Vivamus velit magna, viverra ut commodo et, malesuada ut eros. Donec laoreet erat eget accumsan eleifend. Ut vel venenatis tortor. Maecenas sapien nisi, hendrerit eget velit eu, lobortis dictum massa. Pellentesque maximus et neque vel suscipit. Nunc eget massa sagittis, tincidunt turpis nec, suscipit est. Sed et vulputate massa. Suspendisse a magna sit amet nibh vehicula mattis. Fusce sed porta mauris, vel faucibus nunc. Etiam consectetur tortor eros, nec pretium risus molestie eu."
+  ..summary = "This is a summary for Data Gathering."
   ..content =
       "Data Gathering dolor sit amet, consectetur adipiscing elit. Aenean a mi porttitor, bibendum elit elementum, placerat augue. Quisque eu sollicitudin tortor, sed egestas ante. Sed convallis, mauris quis malesuada convallis, lectus ante vestibulum ante, vel lobortis magna dui eu nisl. Proin ac pellentesque nulla. Morbi facilisis dui aliquam quam pulvinar efficitur. Duis at lorem vitae leo pharetra ultricies. Proin viverra eleifend varius. Nulla sed nisi ut enim placerat venenatis. Maecenas imperdiet accumsan ligula id varius. Donec rhoncus gravida odio vitae convallis.Nullam at tempor erat. Praesent euismod orci nec sollicitudin placerat. Nunc nec nibh efficitur, mattis ante sit amet, scelerisque libero. Aliquam et mollis erat. Pellentesque aliquam convallis turpis sit amet molestie. Duis accumsan venenatis imperdiet. Integer quis est non elit varius mattis. Donec hendrerit in nisl eget sollicitudin. Nulla sapien lacus, mattis non orci sed, commodo tincidunt risus.";
 
 RPConsentSection privacySection = RPConsentSection(RPConsentSectionType.Privacy)
-  ..summary =
-      "This is a summary for Privacy. Lorem ipsum dolor sit amet, consectetur adipiscing elit. Aliquam ut aliquet augue, sed maximus sapien. Cras non sem vulputate, tincidunt est in, euismod orci. Aenean elit nisl, convallis vel lacus eget, pellentesque laoreet nunc. Proin sed turpis odio. Aliquam quis metus eget tortor tristique posuere sed blandit velit. Vivamus velit magna, viverra ut commodo et, malesuada ut eros."
+  ..summary = "This is a summary for Privacy."
   ..content =
       "Privacy dolor sit amet, consectetur adipiscing elit. Aenean a mi porttitor, bibendum elit elementum, placerat augue. Quisque eu sollicitudin tortor, sed egestas ante. Sed convallis, mauris quis malesuada convallis, lectus ante vestibulum ante, vel lobortis magna dui eu nisl. Proin ac pellentesque nulla. Morbi facilisis dui aliquam quam pulvinar efficitur. Duis at lorem vitae leo pharetra ultricies. Proin viverra eleifend varius. Nulla sed nisi ut enim placerat venenatis. Maecenas imperdiet accumsan ligula id varius. Donec rhoncus gravida odio vitae convallis.Nullam at tempor erat. Praesent euismod orci nec sollicitudin placerat. Nunc nec nibh efficitur, mattis ante sit amet, scelerisque libero. Aliquam et mollis erat. Pellentesque aliquam convallis turpis sit amet molestie. Duis accumsan venenatis imperdiet. Integer quis est non elit varius mattis. Donec hendrerit in nisl eget sollicitudin. Nulla sapien lacus, mattis non orci sed, commodo tincidunt risus.";
 
@@ -76,9 +74,8 @@ RPConsentDocument consentDocument = RPConsentDocument('Informed Consent', [
 
 RPConsentReviewStep consentReviewStep =
     RPConsentReviewStep("consentreviewstepID", consentDocument)
-      ..reasonForConsent = 'By tapping AGREE you can take part in the study'
-      ..text = 'Agreed?'
-      ..title = "Consent Review Step Title";
+      ..reasonForConsent = 'informed_consent.agree_text'
+      ..text = 'informed_consent.agree_confirm';
 
 RPVisualConsentStep consentVisualStep =
     RPVisualConsentStep("visualStep", consentDocument);
@@ -90,12 +87,11 @@ RPInstructionStep instructionStep = RPInstructionStep(
       "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Etiam ultricies feugiat turpis nec efficitur. Integer in pharetra libero. Proin a leo eu enim porttitor hendrerit. Suspendisse vestibulum interdum mollis. Donec in sapien ut orci ultricies laoreet. Ut maximus ante id arcu feugiat scelerisque. Proin non rutrum libero. Aliquam blandit arcu ac dolor consequat maximus. Integer et dolor quis quam tempor porta quis vel nibh. Phasellus ullamcorper fringilla lorem, ac tempus sem cursus a. Aliquam maximus facilisis quam. Morbi hendrerit tempor tellus, ac hendrerit augue tincidunt eu. Cras convallis lorem at nulla mattis tristique.",
   footnote: "(1) Important footnote",
   imagePath: "assets/images/waving-hand.png",
-)..text =
-    "Welcome to this study! We would like to kindly ask you to look over these terms and information which gives you an introduction to the study.";
+)..text = "informed_consent.welcome_text";
 
 RPCompletionStep completionStep = RPCompletionStep("completionID")
-  ..title = "Thank You!"
-  ..text = "We saved your consent document.";
+  ..title = "informed_consent.thanks"
+  ..text = "informed_consent.document_saved";
 
 RPOrderedTask consentTask = RPOrderedTask(
   "consentTaskID",