+++ lib/src/model/answerformat/RPChoiceAnswerFormat.dart
@@ -50,8 +50,7 @@ class RPChoice extends Serializable {
     this.value,
     this.isFreeText = false,
     this.detailText,
-  })
-      : super();
+  }) : super();
 
   Function get fromJsonFunction => _$RPChoiceFromJson;
   factory RPChoice.fromJson(Map<String, dynamic> json) =>