+++ lib/src/ui/RPUIFormStep.dart
@@ -158,8 +158,7 @@ class _RPUIFormStepState extends State<RPUIFormStep> {
               padding: const EdgeInsets.all(8.0),
               child: FlatButton(
                 onPressed: () => skipQuestion(),
-                child: Text(RPLocalizations
-                        .of(context)
+                child: Text(RPLocalizations.of(context)
                         .translate("Skip these questions") ??
                     "Skip these questions"),
               ),
@@ -176,8 +175,7 @@ class _RPUIFormStepState extends State<RPUIFormStep> {
           Padding(
             padding: const EdgeInsets.all(8.0),
             child: Text(
-              RPLocalizations
-                      .of(context)
+              RPLocalizations.of(context)
                       ?.translate(widget.formStep.steps[index].title) ??
                   widget.formStep.steps[index].title,
               style: RPStyles.h3,
@@ -219,7 +217,6 @@ class _RPUIFormStepState extends State<RPUIFormStep> {
     );
   }
 
-  @override
   void createAndSendResult() {
     // In this case the result is already created, the only needed thing left is to send it
     blocTask.sendStepResult(stepResult);